add exception

+++ src/main/java/com/senla/hotel/domain/Lodger.java
@@ -2,7 +2,7 @@ package com.senla.hotel.domain;
 
 public class Lodger {
 
-    private Integer id;
+    private Long id;
     private String firstName;
     private String lastName;
     private String phoneNumber;
@@ -16,18 +16,18 @@ public class Lodger {
         this.phoneNumber = phoneNumber;
     }
 
-    public Lodger(Integer id, String firstName, String secondName, String phoneNumber) {
+    public Lodger(Long id, String firstName, String secondName, String phoneNumber) {
         this.id = id;
         this.firstName = firstName;
         this.lastName = secondName;
         this.phoneNumber = phoneNumber;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
