+++ src/main/java/com/senla/hotel/domain/Service.java
@@ -4,7 +4,7 @@ import java.math.BigDecimal;
 
 public class Service {
 
-    private Integer id;
+    private Long id;
     private String name;
     private BigDecimal cost;
 
@@ -16,17 +16,17 @@ public class Service {
         this.cost = cost;
     }
 
-    public Service(Integer id, String name, BigDecimal cost) {
+    public Service(Long id, String name, BigDecimal cost) {
         this.id = id;
         this.name = name;
         this.cost = cost;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
